Moved name-mapping part from main adapter file to a dedicated file. Created lora_common.hpp to hold parts of old lora_adapter.cpp that are reused in the moved parts.
