+++ src/cpp/src/lora_names_mapping.hpp
@@ -4,6 +4,8 @@
 #include <unordered_map>
 #include <set>
 
+#include "lora_common.hpp"
+
 namespace ov {
 namespace genai {
 
@@ -15,6 +17,11 @@ NameMap maybe_map_sgm_blocks_to_diffusers(std::set<std::string> state_dict, int 
 NameMap maybe_map_non_diffusers_lora_to_diffusers(const std::set<std::string>& keys);
 
 void convert_prefix_te(std::string& name);
+
+utils::LoRATensors flux_kohya_lora_preprocessing(const utils::LoRATensors& tensors);
+
+utils::LoRATensors flux_xlabs_lora_preprocessing(const utils::LoRATensors& tensors);
+
 
 }
 }
